push to master

+++ contracts/FixidityLib.sol
@@ -203,7 +203,7 @@ library FixidityLib {
         uint8 decimalDifference;
         if ( _originDigits > _destinationDigits ){
             decimalDifference = _originDigits - _destinationDigits;
-            return x/int256((uint128(10)**uint128(decimalDifference)));
+            return x/int256((uint256(10)**uint256(decimalDifference)));
         }
         else if ( _originDigits < _destinationDigits ){
             decimalDifference = _destinationDigits - _originDigits;
@@ -213,9 +213,9 @@ library FixidityLib {
             //     decimalDifference = abs(_destinationDigits - _originDigits)
             //     decimalDifference < 38
             //     10**38 < 2**128-1
-            assert(x <= maxInt256()/int256(uint128(10)**uint128(decimalDifference)));
-            assert(x >= minInt256()/int256(uint128(10)**uint128(decimalDifference)));
-            return x*(int256(uint128(10)**uint128(decimalDifference)));
+            assert(x <= maxInt256()/int256(uint256(10)**uint256(decimalDifference)));
+            assert(x >= minInt256()/int256(uint256(10)**uint256(decimalDifference)));
+            return x*(int256(uint256(10)**uint256(decimalDifference)));
         }
         // _originDigits == digits()) 
         return x;
